Refactor AleoAmount and Transaction
