+++ dpc/src/testnet1/inner_circuit/inner_circuit.rs
@@ -15,15 +15,15 @@
 // along with the snarkVM library. If not, see <https://www.gnu.org/licenses/>.
 
 use crate::{
-    account::AccountPrivateKey,
     testnet1::{
         inner_circuit_gadget::execute_inner_proof_gadget,
         parameters::SystemParameters,
         record::Record,
         record_encryption::RecordEncryptionGadgetComponents,
-        AleoAmount,
         Testnet1Components,
     },
+    AccountPrivateKey,
+    AleoAmount,
 };
 use snarkvm_algorithms::{
     merkle_tree::{MerklePath, MerkleTreeDigest},
