+++ dpc/src/testnet1/outer_circuit/outer_circuit.rs
@@ -14,12 +14,14 @@
 // You should have received a copy of the GNU General Public License
 // along with the snarkVM library. If not, see <https://www.gnu.org/licenses/>.
 
-use crate::testnet1::{
-    outer_circuit_gadget::execute_outer_proof_gadget,
-    parameters::SystemParameters,
-    program::PrivateProgramInput,
+use crate::{
+    testnet1::{
+        outer_circuit_gadget::execute_outer_proof_gadget,
+        parameters::SystemParameters,
+        program::PrivateProgramInput,
+        Testnet1Components,
+    },
     AleoAmount,
-    Testnet1Components,
 };
 use snarkvm_algorithms::{
     merkle_tree::MerkleTreeDigest,
